remove extra connect function and file
